feat: Add logout functionality
