+++ src/Firebase.js
@@ -1,7 +1,6 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
 import { getAnalytics } from "firebase/analytics";
-import { useNavigate } from 'react-router-dom';
 import {  createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut } from 'firebase/auth';
 import { getAuth } from 'firebase/auth';
 
@@ -26,32 +25,4 @@ const app = initializeApp(firebaseConfig);
 const analytics = getAnalytics(app);
 const auth = getAuth(app); // Initializing the authentication service
 
-//SIGN OUT USER
-const Home = () => {
-    const navigate = useNavigate();
- 
-    const handleLogout = () => {               
-        signOut(auth).then(() => {
-        // Sign-out successful.
-            navigate("/");
-            console.log("Signed out successfully")
-        }).catch((error) => {
-          console.error("Error occured during sign-out:", error);
-        });
-    }
-   
-    return(
-        <>
-          <nav>
-            <p>Welcome Home</p>
-          <div>
-        <button onClick={handleLogout}>
-                        Logout
-              </button>
-        </div>
-          </nav>
-        </>
-    )
-}
- 
-export { app, analytics, auth, getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut , Home }
+export { app, analytics, auth, getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut }
